more refact
