+++ new_structure_outline.md
@@ -4,14 +4,14 @@ Overall new structure of the library
 
 pyowc/core
     utils.py: change_dates, cal2dec, potential_vorticity, wrap_longitudes, sorter
-    stats.py: brk_pt_fit, build_cov, covarxy_pv, calc_piecewisefit, covar_xyt_pv, noise_variance, signal_variance, spatial_correlation
+    stats.py: brk_pt_fit, build_cov, covarxy_pv, calc_piecewisefit, covar_xyt_pv, noise_variance, signal_variance, spatial_correlation, fit_cond
     finders.py: find_10thetas, find_25boxes, find_besthit, find_ellipse
 
 pyowc/data
     fetchers.py: get_region_data, get_region_hist_locations, get_data, get_topo_grid
     wrangling.py: interp_climatology, map_data_grid 
 
-pyowc/calibration.py: fit_cond, nlbpfun, set_calseries, update_salinity_mapping
+pyowc/calibration.py: set_calseries, update_salinity_mapping
 
 pyowc/configuration.py: load_configuration
