+++ pyowc/__init__.py
@@ -1,10 +1,11 @@
 """ Global import controler
 """
 
+#pylint: disable=broad-except
 try:
     import pkg_resources
     __version__ = pkg_resources.get_distribution("pyowc").version
-except ModuleNotFoundError:
+except Exception:
     # Local copy, not installed with setuptools, or setuptools is not available.
     # Disable minimum version checks on downstream libraries.
     __version__ = "999"
@@ -12,6 +13,7 @@ except ModuleNotFoundError:
 from . import core
 from . import data
 from . import plot
+from . import utilities
 from .plot import dashboard
 
 from . import calibration
@@ -27,6 +29,7 @@ __all__ = (
     "data",
     "plot",
     "dashboard",
+    "utilities",
     # Constants
     "__version__"
 )
