+++ pyowc/tests/potential_vorticity_test.py
@@ -13,7 +13,7 @@ https://gitlab.noc.soton.ac.uk/edsmall/bodc-dmqc-python
 
 import unittest
 import numpy as np
-from pyowc.core.utils import potential_vorticity
+from pyowc.utilities import potential_vorticity
 
 class PotentialVorticityTestCase(unittest.TestCase):
     """
